Use www subdomain instead of apex domain
Apex domains configured with A-records can cause a number of technical
problems related to DNS [1], and performance and security concerns
related to cookies [2]. The advantages of using the apex domain are
primarily aesthetic. I have concluded that it is best practice to use
www.example.com as the canonical domain name instead of example.com.

  [1]: https://devcenter.heroku.com/articles/avoiding-naked-domains-dns-arecords
  [2]: http://developer.yahoo.com/performance/rules.html#cookie_free